// frontend/src/api-client/model/document-dto.ts
/**
 * Inkweld API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DocumentDto { 
    /**
     * Unique identifier for the document
     */
    id: string;
    /**
     * The name of the document
     */
    name: string;
    /**
     * The user ID of the document owner
     */
    ownerId: string;
    /**
     * The username part of the document ID
     */
    username: string;
    /**
     * The project slug part of the document ID
     */
    projectSlug: string;
    /**
     * Timestamp when the document was last modified
     */
    lastModified: string;
}